* Better portability of recently comitted IPv6 code - Thanks Antonio Querubin

+++ src/policyd.h
@@ -232,7 +232,7 @@ unsigned long int mysql_timeout;        /* mysql query timeout   */
   int    cidr_ip_match (unsigned long ip, char *range);
   pid_t w_fork(void);
  const char *w_inet_ntop(struct sockaddr *addrptr, char *strptr, size_t len);
- uint16_t sockaddrport(struct sockaddr *addrptr);
+ int sockaddrport(struct sockaddr *addrptr);
  ssize_t w_read(unsigned int fd, char *ptr, size_t max_size);
  ssize_t w_write(unsigned int fd, const void *vbuf);
  ssize_t f_write(unsigned int volatile fd, const void *vptr, size_t n);
